Refactor player leaderboard to extract player table row
